CSS Fixes
Fixing the HTML body when on smaller devices. There was a lot of white
space on the right that cause a horizontal scroll.

Fixed the navbar anchor tag on hover to be the same as the rest of the
page for consistency. The lovely animation to change them to red when
hovered.