Improving logo text & adding js for hiding/showing text
The logo for the site is just going to be text for now. Created some
media queries to help scale the text properly.

Since "Iron Technologies" also shows up in the navigation bar. Created
a simple script that will detect when the main logo contain is in the
viewport. When it's in the viewport the text in the navigation bar will
be hidden. Once the main logo is out of the view port then the text in
the navbar will display.

This is to reduce redundant text that could confuse the user.